refactor: extract EngagementService from DispatchOperations
First lane peeled off the DispatchOperations god object, over the
LaunchAdmission guard seam (#206). Room engagement launches nothing and
reserves no container — it is pure room state (one JSON value on the spec
row) plus an optional engage-time snapshot — so it has no business living
in the dispatch launcher.

Move engage/completeEngage/disengage and their helpers into an
EngagementService that collaborates only with SpecStore, ProjectLoader,
LaunchAdmission, the event sink, and (for the snapshot) the shell — no run
launch, no reservation, no run rows. DispatchOperations holds one
EngagementService and its engage/disengage stay as thin delegators, so the
public surface and callers are unchanged. Dropped the dead `actor` param
persistEngagement never used. DispatchOperations sheds ~130 lines.

Isolating engagement here front-loads the room/spec decouple: the room's
collaboration surface can evolve without touching dispatch.

TDD: EngagementService is fully unit-tested through the unchanged
EngagementLifecyclet harness, plus three added cases (engage/disengage a
missing spec → SPEC_NOT_FOUND; a spec deleted mid-snapshot → engage_failed)
that cover its defensive not-found paths. mvn clean verify green, api.*
100% coverage held.
